判別さんのテストと実装
テストファイルのhanbetu.phpを消してhanbetuTest.phpに統合した。
