+++ tests/hanbetuTest.php
@@ -1,15 +1,17 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: y-murakami
- * Date: 2015/07/10
- * Time: 22:59
- */
-
+require_once("../applications/hanbetu.php");
 class hanbetuTest extends PHPUnit_Framework_TestCase
 {
-	public function testFuga()
+	public function test_usable()
 	{
+		$hanbetu = new hanbetu();
+		$this->assertTrue($hanbetu->is_usable(new tenyen()));
+	}
 
+	public function test_not_usable()
+	{
+		$hanbetu = new hanbetu();
+		$this->assertFalse($hanbetu->is_usable(new oneyen()));
+		$this->assertFalse($hanbetu->is_usable(new fiveyen()));
 	}
 }
